feat: DI with Spring annotation
