+++ src/main/java/org/example/controller/PostController.java
@@ -3,11 +3,13 @@ package org.example.controller;
 import com.google.gson.Gson;
 import org.example.model.Post;
 import org.example.service.PostService;
+import org.springframework.stereotype.Controller;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.Reader;
 
+@Controller
 public class PostController {
   public static final String APPLICATION_JSON = "application/json";
   private final PostService service;
