+++ src/main/java/org/example/servlet/MainServlet.java
@@ -2,7 +2,8 @@ package org.example.servlet;
 
 import org.example.controller.PostController;
 import org.example.service.PostService;
-import org.example.repository.PostRepository;
+import org.example.repository.PostRepositoryStubImpl;
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -19,9 +20,8 @@ public class MainServlet extends HttpServlet {
 
   @Override
   public void init() {
-    final var repository = new PostRepository();
-    final var service = new PostService(repository);
-    controller = new PostController(service);
+    final var context = new AnnotationConfigApplicationContext("org.example");
+    controller = context.getBean("postController", PostController.class);
   }
 
   @Override
